Many changes (see desc)
Allowed the loader config to be optional (as was intended), added an error for invalid library types in config, added debug library load logging message, changed late service adding to comply with new way in Base, add functionality so services can be unloaded & loaded.

This is not completely extensive (if you can believe it) but a key note is that library unloading is BUGGED if you use the automatic loader.